items-loadgen/requirements.txt
Faker==23.0.0
psycopg2-binary==2.9.9